+++ SOUP.user.js
@@ -3,7 +3,7 @@
 // @namespace   https://github.com/vyznev/
 // @description Miscellaneous client-side fixes for bugs on Stack Exchange sites (development)
 // @author      Ilmari Karonen
-// @version     1.47.13
+// @version     1.47.14
 // @copyright   2014-2017, Ilmari Karonen (https://stackapps.com/users/10283/ilmari-karonen)
 // @license     ISC; https://opensource.org/licenses/ISC
 // @match       *://*.stackexchange.com/*
@@ -1485,7 +1485,39 @@ fixes.mse295065 = {
 		} );
 	}
 };
+fixes.mso345590 = {
+	title:	"The Stack Exchange menu is partly covered by the scrollbar when the window is too narrow",
+	url:	"https://meta.stackoverflow.com/q/345590",
+	script:	function () {
+		var $window = $(window), $header = $('.js-so-header');
+		if ( $header.length != 1 ) return;
 
+		// override unwanted .so-header._fixed{ max-width: auto } style
+		$header.css( 'min-width', $('#content').outerWidth() );
+
+		// based on https://stackoverflow.com/a/12958987
+		function scrollHeader () { $header.css( 'left', -$window.scrollLeft() ) }
+
+		var isActive = false;
+		function maybeToggleHeaderFix () {
+			var isFixed = $header.hasClass('_fixed');
+			if ( isFixed === isActive ) return;
+			if ( isFixed ) {
+				$window.on( 'scroll resize', scrollHeader );
+				scrollHeader();
+			} else {
+				$window.off( 'scroll resize', scrollHeader );
+				$header.css( 'left', 0 );
+			}
+			isActive = isFixed;
+		}
+		// set up an observer in case the _fixed class is dynamically added or removed
+		var observer = new MutationObserver( maybeToggleHeaderFix );
+		observer.observe( $header[0], { attributes: true } );
+
+		maybeToggleHeaderFix();
+	}
+};
 
 
 //